requirements.txt
autopep8==1.6.0
bidict==0.22.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.2.2
idna==3.3
image-quality @ git+https://github.com/ocampor/image-quality@a0435e2966883b3420ee8be16a3ef9c6a8de79aa
imageio==2.25.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
libsvm==3.23.0.4
MarkupSafe==2.1.2
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
pycodestyle==2.8.0
python-engineio==4.3.3
python-socketio==5.7.1
PyWavelets==1.4.1
requests==2.28.1
scikit-image==0.19.3
scipy==1.10.0
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
tifffile==2023.2.3
toml==0.10.2
urllib3==1.26.10
websocket-client==1.3.3
Werkzeug==2.2.2
zipp==3.12.1
